% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.cumsum.R
\name{beta.cumsum}
\alias{beta.cumsum}
\title{Médias e ICs das probabilidades de notificação cumulativa a cada dia}
\usage{
beta.cumsum(NobBS.output, NobBS.params.post, samples)
}
\arguments{
\item{NobBS.output}{objeto retornado pela função NobBS do pacote de
mesmo nome Este argumento é ignorado se o argumento
NobBS.params.post é usado}

\item{NobBS.params.post}{data frame com as distribuicoes
posteriores dos parâmetros estimados pela função NobBS. Está
contido na lista que é retornada pela função}

\item{samples}{Número de samples a ser considerado}
}
\value{
data frame com média e quantis 2.5\% e 97.5\% das
distribuições a posteriori dos parâmetros de atraso de
notificação pelo método de nowcasting da função NobBS. Os
valores estão convertidos para escala de probabilidade são cumulativos, e
portanto podem ser interpretados como a probabilidade cumulativa de casos
ser notificados D dias após o dias o primeiro sintoma, sendo que
vai de zero ao máximo definido pelos argumentos do nowcasting
}
\description{
Médias e ICs das probabilidades de notificação cumulativa a cada dia
}
