% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.summary.R
\name{beta.summary}
\alias{beta.summary}
\title{Médias e ICs das probabilidades de notificação a cada dia}
\usage{
beta.summary(NobBS.output, NobBS.params.post)
}
\arguments{
\item{NobBS.output}{objeto retornado pela função NobBS do pacote de
mesmo nome Este argumento é ignorado se o argumento
NobBS.params.post é usado.}

\item{NobBS.params.post}{data frame com as distribuicoes
posteriores dos parâmetros estimados pela função NobBS. Está
contido na lista que é retornada pela função.}
}
\value{
data frame com média e quantis 2.5\% e 97.5\% das
distribuições a posteriori dos parâmetros de atraso de
notificação pelo método de nowcasting da função NobBS. Os
valores estão convertidos para escala de probabilidade, e
portanto podem ser interpretado como a probabilidade de um caso
ser notificado D dias após o dias o primeiro sintoma, sendo que
vai de zero ao máximo definido pelos argumentos do nowcasting
}
\description{
Médias e ICs das probabilidades de notificação a cada dia
}
