% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.lines.R
\name{count.lines}
\alias{count.lines}
\title{Conta número de linhas de um arquivo de texto, zipado ou bzipado ou xzeado}
\usage{
count.lines(filename)
}
\arguments{
\item{filename}{Caractere. Caminho do arquivo}
}
\description{
Conta número de linhas de um arquivo de texto, zipado ou bzipado ou xzeado
}
