% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countByAgeClass.R
\name{countByAgeClass}
\alias{countByAgeClass}
\title{countByAgeClass}
\usage{
countByAgeClass(df, age_table = age_table)
}
\arguments{
\item{df}{data.frame}

\item{age_table}{age table}
}
\description{
countByAgeClass
}
