% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt.rw.R
\name{dt.rw}
\alias{dt.rw}
\title{Tempo de duplicação ao longo de uma serie temporal de n de casos}
\usage{
dt.rw(zoo.obj, window.width)
}
\arguments{
\item{zoo.obj}{objeto da classe 'zoo' com uma serie temporal
univariada (n de casos)}

\item{window.width}{largura da janela (em unidades de tempo da
serie temporal, em geral dias)}
}
\description{
Tempo de duplicação ao longo de uma serie temporal de n de casos
}
\details{
Toma uma série temporal de n de casos, ajusta a função
    fitP.exp a 'running windows' de n de dias fixo ao longo da
    série. Retorna um objeto de série temporal com os valores dos
    tempos de duplicação em cada janela e seus intervalos de
    confiança
}
