% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estima.not.R
\name{estima.not}
\alias{estima.not}
\title{Estima numero de notificacoes por dia a partir de um vetor de n de
casos novos e da distribuição de probabilidades de notificação do
nowcasting}
\usage{
estima.not(
  vetor.casos,
  NobBS.output,
  NobBS.params.post,
  from = length(vetor.casos) - 30
)
}
\arguments{
\item{vetor.casos}{objeto da classe zoo com n de casos}

\item{NobBS.output}{objeto retornado pela função NobBS do pacote de
mesmo nome. Este argumento é ignorado se o argumento
NobBS.params.post é usado.}

\item{NobBS.params.post}{data frame com as distribuicoes
posteriores dos parâmetros estimados pela função NobBS. Está
contido na lista que é retornada pela função.}

\item{from}{posicao do vetor de casos a partir da qual estimar o
numero de notificacões}
}
\description{
Estima numero de notificacoes por dia a partir de um vetor de n de
casos novos e da distribuição de probabilidades de notificação do
nowcasting
}
