% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/existe.nowcasting.R
\name{existe.nowcasting}
\alias{existe.nowcasting}
\title{Função para checar se existem dados de nowcasting para a unidade administrativa}
\usage{
existe.nowcasting(tipo, data, output.dir)
}
\arguments{
\item{tipo}{covid, srag, obitos_covid, obitos_srag}

\item{data}{data}

\item{output.dir}{diretorio}
}
\description{
Função para checar se existem dados de nowcasting para a unidade administrativa
}
