% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNowcastedLines.R
\name{fillNowcastedLines}
\alias{fillNowcastedLines}
\title{Using Nowcast and survival analysis to complete covid table with unobserved cases}
\usage{
fillNowcastedLines(
  df,
  nowcast,
  hosp_wait_fit,
  int_wait_fit,
  UTI_stay_wait_fit,
  UTI_after_wait_fit,
  prob_UTI,
  ...
)
}
\arguments{
\item{df}{foo}

\item{nowcast}{foo}

\item{hosp_wait_fit}{foo}

\item{int_wait_fit}{foo}

\item{UTI_stay_wait_fit}{foo}

\item{UTI_after_wait_fit}{foo}

\item{prob_UTI}{foo}

\item{...}{}
}
\value{

}
\description{
Using Nowcast and survival analysis to complete covid table with unobserved cases
}
