% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillNowcastedLines.R
\name{fillNowcastedLinesFast}
\alias{fillNowcastedLinesFast}
\title{fillNowcastedLinesFast}
\usage{
fillNowcastedLinesFast(
  df,
  nowcast,
  hosp_wait_fit,
  int_wait_fit,
  UTI_stay_wait_fit,
  UTI_after_wait_fit,
  prob_UTI,
  ...
)
}
\arguments{
\item{df}{df}

\item{nowcast}{nowcast}

\item{hosp_wait_fit}{hosp_wait_fit}

\item{int_wait_fit}{int_wait_fit}

\item{UTI_stay_wait_fit}{UTI_stay_wait_fit}

\item{UTI_after_wait_fit}{UTI_after_wait_fit}

\item{prob_UTI}{prob_UTI}

\item{...}{...}
}
\description{
fillNowcastedLinesFast
}
