% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.esusve.R
\name{filtros.esusve}
\alias{filtros.esusve}
\title{Title}
\usage{
filtros.esusve(dados, tipo, remove.col = TRUE)
}
\arguments{
\item{dados}{df}

\item{tipo}{covid, srag, obitos_covid, obitos_srag}

\item{remove.col}{Lógico. Se TRUE seleciona apenas colunas de datas usadas em nowcasting}
}
\description{
Title
}
