% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.sivep.R
\name{filtros.sivep}
\alias{filtros.sivep}
\title{Title}
\usage{
filtros.sivep(dados, tipo, hospitalizados)
}
\arguments{
\item{dados}{df}

\item{tipo}{covid, srag, obitos_covid, obitos_srag}

\item{hospitalizados}{Logical. If TRUE filters only hosptalized cases}
}
\description{
Title
}
