% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitP.exp.R
\name{fitP.exp}
\alias{fitP.exp}
\title{Ajusta um modelo exponencial para n de casos em função de n de dias
passados desde o início da série.}
\usage{
fitP.exp(zoo.obj, only.coef = TRUE)
}
\arguments{
\item{zoo.obj}{objeto da classe zoo com série temporal univariada}

\item{only.coef}{se TRUE a função retorna um vetor coms
coeficientes da regressão e seus intervalos de confiança. Se
FALSE retorna o objeto do modelo ajustado, da classe glm.}
}
\description{
Ajusta um modelo exponencial para n de casos em função de n de dias
passados desde o início da série.
}
\details{
Ajusta um modelo generalizado misto para contagens
(Poisson) dos valores de uma série temporal em função do número
de dias transcorridos desde o início da série. Como o modelo
Poisson tem função de ligação log, equivale a ajustar uma
função de crescimento exponencial às contagens, mas com erro
não gaussianos e sim Poisson.
}
