% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.exponential.R
\name{forecast.exponential}
\alias{forecast.exponential}
\title{Forecast usando regressão Poisson sobre série dos casos acumulados}
\usage{
forecast.exponential(zoo.obj, start, end = length(zoo.obj), days.forecast, ...)
}
\description{
Forecast usando regressão Poisson sobre série dos casos acumulados
}
