% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formata.now.df.R
\name{formata.now.df}
\alias{formata.now.df}
\title{Funcao para formatar data.frame para o grafico de nowcasting casos diarios e acumulados}
\usage{
formata.now.df(now.pred.zoo, now.proj.zoo, lista)
}
\arguments{
\item{now.pred.zoo}{previstos}

\item{now.proj.zoo}{projetado}

\item{lista}{lista}
}
\description{
Funcao para formatar data.frame para o grafico de nowcasting casos diarios e acumulados
}
