% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera.nowcasting.R
\name{gera.nowcasting}
\alias{gera.nowcasting}
\title{Função para gerar nowcasting}
\usage{
gera.nowcasting(
  dados,
  caso = TRUE,
  tipo,
  hospitalizados = TRUE,
  trim.now,
  window,
  trajectories = FALSE
)
}
\arguments{
\item{dados}{dados}

\item{caso}{Lógico. Se TRUE faz casos e FALSE, óbitos}

\item{tipo}{covid ou srag}

\item{hospitalizados}{Lógico, filtra apenas hospitalizados}

\item{trim.now}{corte para nowcasting}

\item{window}{janela para nowcasting}

\item{trajectories}{Lógico. Retorna trajetórias}
}
\description{
Função para gerar nowcasting
}
