% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.base.R
\name{get.data.base}
\alias{get.data.base}
\title{Extrai as datas de um vetor (para pegar a última data ver get.last.date)}
\usage{
get.data.base(names, last = FALSE)
}
\arguments{
\item{names}{vector}

\item{last}{Logical, whether the most recent date of the vector should be returned}
}
\description{
Extrai as datas de um vetor (para pegar a última data ver get.last.date)
}
