% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.base2.R
\name{get.data.base2}
\alias{get.data.base2}
\title{Extrai a data mais recente de nowcasting por localidade}
\usage{
get.data.base2(adm, sigla.adm, tipo)
}
\arguments{
\item{adm}{unidade administrativa}

\item{sigla.adm}{sigla da unidade administrativa}

\item{tipo}{covid, srag, obitos_covid, obitos_srag}
}
\description{
Extrai a data mais recente de nowcasting por localidade
}
