% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.last.date.R
\name{get.last.date}
\alias{get.last.date}
\title{Pega a data mais recente da base de dados}
\usage{
get.last.date(dir)
}
\arguments{
\item{dir}{diretorio}
}
\value{

}
\description{
Pega a data mais recente da base de dados
}
