% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCurrentInBed.R
\name{getCurrentInBed}
\alias{getCurrentInBed}
\title{getCurrentInBed}
\usage{
getCurrentInBed(df, date, UTI)
}
\arguments{
\item{df}{df}

\item{date}{date}

\item{UTI}{UTI}
}
\description{
getCurrentInBed
}
