% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.hospital.probabilities.R
\name{getProbDeath}
\alias{getProbDeath}
\title{Probabilidade de morte de hospitalizado comum, e em UTI,}
\usage{
getProbDeath(df, UTI = FALSE)
}
\arguments{
\item{df}{df}

\item{UTI}{UTI}
}
\description{
Probabilidade de morte de hospitalizado comum, e em UTI,
}
