% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.times.R
\name{getTimes}
\alias{getTimes}
\title{getTimes}
\usage{
getTimes(x, late, early, censored = FALSE)
}
\arguments{
\item{x}{x}

\item{late}{late}

\item{early}{early}

\item{censored}{censored}
}
\description{
getTimes
}
