% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHospitalTable.R
\name{makeHospitalTable}
\alias{makeHospitalTable}
\title{makeHospitalTable}
\usage{
makeHospitalTable(df.table, dates, UTI = FALSE)
}
\arguments{
\item{df.table}{df.table}

\item{dates}{dates}

\item{UTI}{UTI}
}
\description{
makeHospitalTable
}
