% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NobBS.posterior.R
\name{NobBS.posterior}
\alias{NobBS.posterior}
\title{Title}
\usage{
NobBS.posterior(
  data,
  now,
  units,
  onset_date,
  report_date,
  moving_window = NULL,
  max_D = NULL,
  cutoff_D = NULL,
  proportion_reported = 1,
  quiet = TRUE,
  specs = list(dist = c("Poisson", "NB"), alpha1.mean.prior = 0, alpha1.prec.prior =
    0.001, alphat.shape.prior = 0.001, alphat.rate.prior = 0.001, beta.priors = NULL,
    param_names = NULL, conf = 0.95, dispersion.prior = NULL, nAdapt = 1000, nChains = 1,
    nBurnin = 1000, nThin = 1, nSamp = 10000)
)
}
\arguments{
\item{specs}{}
}
\value{

}
\description{
Title
}
