% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now.proj.R
\name{now.proj}
\alias{now.proj}
\title{Funcao para projecao do n acumulado de casos por data do sintoma}
\usage{
now.proj(pred, pred.original, now.params.post, n.dias = 5, data)
}
\arguments{
\item{pred}{Data frame. Objeto \code{now.pred.zoo} gerado em prepara_nowcasting.R}

\item{pred.original}{Data frame. Objeto \code{now.pred.original} gerado em prepara_nowcasting.R}

\item{now.params.post}{Data frame. Objeto \code{now.params.post} gerado em gerado em prepara_dados_nowcasting.R}
}
\description{
Funcao para projecao do n acumulado de casos por data do sintoma
}
