% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.cumsum.R
\name{nowcasting.cumsum}
\alias{nowcasting.cumsum}
\title{Médias e ICs das estimativas de nowcasting cumulativa a cada dia}
\usage{
nowcasting.cumsum(
  NobBS.posterior.output,
  NobBS.posterior.output.trajectories,
  samples
)
}
\arguments{
\item{NobBS.posterior.output}{objeto retornado pela função NobBS.posterior.}

\item{NobBS.posterior.output.trajectories}{trajetórias das estimativas de nowcasting a posteriori,
utilizadas no calculo de estimativas cumulativas}

\item{samples}{Número de samples a ser considerado.}
}
\value{
data frame com média e quantis 2.5\% e 97.5\% das
distribuições a posteriori das estimativas de nowcasting da função NobBS.posterior.
Os valores são cumulativos, e portanto podem ser interpretados como a estimativa cumulativa de casos
ser notificados D dias após o dias o primeiro sintoma, sendo que
vai de zero ao máximo definido pelos argumentos do nowcasting
}
\description{
Médias e ICs das estimativas de nowcasting cumulativa a cada dia
}
