% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.estimate.R0.R
\name{plot.estimate.R0}
\alias{plot.estimate.R0}
\title{Funcao para plot do R efetivo}
\usage{
\method{plot}{estimate.R0}(df.re, ...)
}
\arguments{
\item{df.re}{df.re}

\item{...}{...}
}
\description{
Funcao para plot do R efetivo
}
