% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nowcast.acumulado.R
\name{plot.nowcast.acumulado}
\alias{plot.nowcast.acumulado}
\title{Funcao para fazer o plot de nowcasting acumulado, com valores estimados e notificados, com projecao para os proximos 5 dias}
\usage{
\method{plot}{nowcast.acumulado}(df)
}
\arguments{
\item{df}{}
}
\description{
Funcao para fazer o plot de nowcasting acumulado, com valores estimados e notificados, com projecao para os proximos 5 dias
}
