% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nowcast.diario.R
\name{plot.nowcast.diario}
\alias{plot.nowcast.diario}
\title{Funcao para fazer plot nowcast de casos diarios (notificado e nowcasting), com linha de tendencia atenuada e IC}
\usage{
\method{plot}{nowcast.diario}(df)
}
\arguments{
\item{df}{}
}
\description{
Funcao para fazer plot nowcast de casos diarios (notificado e nowcasting), com linha de tendencia atenuada e IC
}
