% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEstimateR0.R
\name{plotEstimateR0}
\alias{plotEstimateR0}
\title{Funcao para plot do R efetivo}
\usage{
plotEstimateR0(x, ...)
}
\arguments{
\item{x}{df.re}

\item{...}{...}
}
\description{
Funcao para plot do R efetivo
}
