% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNowcastAcumulado.R
\name{plotNowcastAcumulado}
\alias{plotNowcastAcumulado}
\title{Funcao para fazer o plot de nowcasting acumulado, com valores estimados e notificados, com projecao para os proximos 5 dias}
\usage{
plotNowcastAcumulado(x, ...)
}
\arguments{
\item{x}{df}

\item{...}{...}
}
\description{
Funcao para fazer o plot de nowcasting acumulado, com valores estimados e notificados, com projecao para os proximos 5 dias
}
