% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNowcastDiario.R
\name{plotNowcastDiario}
\alias{plotNowcastDiario}
\title{Funcao para fazer plot nowcast de casos diarios (notificado e nowcasting), com linha de tendencia atenuada e IC}
\usage{
plotNowcastDiario(x, ...)
}
\arguments{
\item{x}{df}

\item{...}{...}
}
\description{
Funcao para fazer plot nowcast de casos diarios (notificado e nowcasting), com linha de tendencia atenuada e IC
}
