% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{posteriors}
\alias{posteriors}
\title{Estimate posterior probabilities for nowcasting AND SI variation}
\usage{
posteriors(data)
}
\arguments{
\item{data}{data}
}
\description{
Estimate posterior probabilities for nowcasting AND SI variation
}
