% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preenche.now.R
\name{preenche.now}
\alias{preenche.now}
\title{Preenche NA's iniciais do vetor de estimado pelo nowcasting}
\usage{
preenche.now(vetor.now, vetor.casos)
}
\arguments{
\item{vetor.now}{vetor com número de casos estimados pelo
nowcasting, com NAs nas datas para as quais não há estimativas.}

\item{vetor.casos}{vetor com numero de casos o observados. Deve ter
mesmo comprimento de 'vetor.now'}
}
\description{
Preenche NA's iniciais do vetor de estimado pelo nowcasting
}
\details{
O nowcasting retornado pela função NobBS usado com janela
(argumento 'moving_window') ou limite máximo de atraso
(argumento 'max_D') produz estimativas para os últimos dias,
definidos por esses argumentos. Esta função preenche os dias
anteriores com os valores de um outro vetor, normalmente o
vetor de n de casos observado
}
