% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Re.com.data.R
\name{Re.com.data}
\alias{Re.com.data}
\title{Calcula R efetivo sobre a estimativas de nowcasting retornadas pela função NobBS::NobBS}
\usage{
\method{Re}{com.data}(ncasos, datas, dia0 = min(datas), delay = 5)
}
\arguments{
\item{ncasos}{vetor de número de novos casos}

\item{datas}{vetor de datas dos novos casos}

\item{dia0}{dia zero}

\item{delay}{atraso}
}
\description{
Calcula R efetivo sobre a estimativas de nowcasting retornadas pela função NobBS::NobBS
}
