% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Re.nowcasting.R
\name{re.nowcasting}
\alias{re.nowcasting}
\title{Calcula R efetivo sobre as trajetórias de nowcasting retornadas pela função NobBS.posterior}
\usage{
re.nowcasting(R.method, trajectories, Nsamples, quantiles = TRUE, ...)
}
\arguments{
\item{R.method}{função parcialmente aplicada que calcula Re de uma trajetória}

\item{trajectories}{data.frame com trajetórias de nowcasting}

\item{Nsamples}{int número de amostrar das trajetórias de nowcasting}

\item{quantiles}{logical retorna quantis da distribuição posterior de R ou as osteriores completas?}

\item{...}{argumentos extras são repassados pro ldply: .parallel=T é especialmente útil}
}
\description{
Calcula R efetivo sobre as trajetórias de nowcasting retornadas pela função NobBS.posterior
}
