% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.esus.generica.R
\name{read.esus.generica}
\alias{read.esus.generica}
\title{Função para ler base da eSUS-VE}
\usage{
read.esus.generica(file.name, encoding = NULL, ...)
}
\arguments{
\item{file.name}{file name}

\item{encoding}{encoding}

\item{...}{...}
}
\description{
Função para ler base da eSUS-VE
}
