% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sivep.generica.R
\name{read.sivep.generica}
\alias{read.sivep.generica}
\title{Função para ler base da sivep}
\usage{
read.sivep.generica(file.name, ...)
}
\arguments{
\item{file.name}{file name}

\item{...}{...}
}
\description{
Função para ler base da sivep
}
