% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.esus.R, R/read.sivep.R
\name{read.sivep}
\alias{read.sivep}
\title{Função para ler base da sivep}
\usage{
read.sivep(dir, escala, sigla, geocode, data, residentes = TRUE, ...)

read.sivep(dir, escala, sigla, geocode, data, residentes = TRUE, ...)
}
\arguments{
\item{dir}{Caractere. Caminho relativo para o diretório onde está o dado}

\item{escala}{Caractere. Escala de análise aceita: `"pais"`, `"estado"`, `"municipio"` #ö micro e meso um dia}

\item{geocode}{Caractere. Geocode IBGE do estado ou município. Município pde ter 6 ou 7 dígitos}

\item{data}{Caractere. Data no formato  "%Y_%m_%d". Quando NULL (padrão) pega a data mais recente}

\item{residentes}{Filtar por residentes ou por local da notificação}

\item{...}{Qualquer parâmetro de `read.csv()`}
}
\description{
Função para ler base da sivep

Função para ler base da sivep
}
