% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_times.R
\name{rwaittime_posterior_age}
\alias{rwaittime_posterior_age}
\title{rwaittime_posterior_age}
\usage{
rwaittime_posterior_age(n, age, fit)
}
\arguments{
\item{n}{n}

\item{age}{age}

\item{fit}{fit}
}
\description{
rwaittime_posterior_age
}
