% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabelas.web.R
\name{tabelas.web}
\alias{tabelas.web}
\title{Title}
\usage{
tabelas.web(output.dir, tipo, df.cum, df.td, df.re = NULL, data_base)
}
\arguments{
\item{output.dir}{output.dir}

\item{tipo}{tipo}

\item{df.cum}{df.cum}

\item{df.td}{df.td}

\item{df.re}{df.re}

\item{data_base}{data_base}
}
\description{
Title
}
