% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themeObservatorio.R
\name{themeObservatorio}
\alias{themeObservatorio}
\title{Theme to use following Observatorio COVID BR website standards}
\usage{
themeObservatorio()
}
\description{
Theme to use following Observatorio COVID BR website standards
}
