% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diazero.R
\name{diazero}
\alias{diazero}
\title{Corta uma série temporal no dia zero.}
\usage{
diazero(zoo.obj, limite)
}
\arguments{
\item{zoo.obj}{objeto da classe zoo}

\item{limite}{limite para corte}
}
\description{
Corta uma série temporal no dia zero.
}
\details{
Esta função corta um objeto da classe zoo, tirando todos
os valores anteriores ao primeiro valor igual a um certo limite
(n.casos).
}
