% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R0.R
\name{estimate.R0}
\alias{estimate.R0}
\title{Estima R efetivo}
\usage{
estimate.R0(novos.casos, day0 = 8, delay = 7, ...)
}
\arguments{
\item{novos.casos}{vetor com novos casos}

\item{day0}{dia zero}

\item{delay}{atraso}

\item{...}{outros parâmetros}
}
\description{
Estima R efetivo
}
