% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtros.sivep.R
\name{filtros.sivep}
\alias{filtros.sivep}
\title{Title}
\usage{
filtros.sivep(dados, tipo, hospitalizados, remove.col)
}
\arguments{
\item{dados}{df}

\item{tipo}{covid, srag, obitos_covid, obitos_srag}

\item{hospitalizados}{Lógico. Se TRUE filtra apenas casos hospitalizados}

\item{remove.col}{Lógico. Se TRUE seleciona apenas colunas de datas usadas em nowcasting}
}
\description{
Title
}
