% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.exponential.R
\name{forecast.exponential}
\alias{forecast.exponential}
\title{Forecast usando regressão Poisson sobre série dos casos acumulados}
\usage{
forecast.exponential(zoo.obj, start, end = length(zoo.obj), days.forecast, ...)
}
\arguments{
\item{zoo.obj}{objeto zoo}

\item{start}{inicio}

\item{end}{fim}

\item{days.forecast}{datas previsão}

\item{...}{...}
}
\description{
Forecast usando regressão Poisson sobre série dos casos acumulados
}
