% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.base.R
\name{get.data.base}
\alias{get.data.base}
\title{Extrai a data mais recente de nowcasting}
\usage{
get.data.base(names)
}
\arguments{
\item{names}{vector}
}
\description{
Extrai a data mais recente de nowcasting
}
