% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.times.R
\name{getTimes}
\alias{getTimes}
\title{Title}
\usage{
getTimes(x, late, early, censored = FALSE)
}
\arguments{
\item{x}{foo}

\item{late}{foo}

\item{early}{foo}

\item{censored}{foo}
}
\value{

}
\description{
Title
}
