% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowfcts-package.R
\docType{package}
\name{nowfcts-package}
\alias{nowfcts}
\alias{nowfcts-package}
\title{nowfcts: Auxiliary functions for analyzing COVID-19 data from different sources}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: SRAG Taskforce \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
