% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimesValidation.R
\name{plotTimesValidation}
\alias{plotTimesValidation}
\title{plotTimesValidation}
\usage{
plotTimesValidation(times_table, fit1, age = TRUE)
}
\arguments{
\item{age}{}
}
\description{
plotTimesValidation
}
