% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.notificacoes.data.R
\name{write.notificacoes.data}
\alias{write.notificacoes.data}
\title{Formata as tabelas de numero de casos por data de notificacao e sintomas}
\usage{
write.notificacoes.data(dados, output.dir, tipo = "covid", data)
}
\arguments{
\item{dados}{dados}

\item{output.dir}{output.dir}

\item{tipo}{tipo}

\item{data}{data}
}
\description{
Formata as tabelas de numero de casos por data de notificacao e sintomas
}
