% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.notificacoes.data.R
\name{write.notificacoes.data}
\alias{write.notificacoes.data}
\title{Title}
\usage{
write.notificacoes.data(dados, output.dir, tipo = "covid", data)
}
\arguments{
\item{data}{}
}
\description{
Title
}
