% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.nowcasting.R
\name{write.nowcasting}
\alias{write.nowcasting}
\title{Writes nowcasting}
\usage{
write.nowcasting(now, output.dir, tipo = "covid", data)
}
\arguments{
\item{data}{}
}
\description{
Writes nowcasting
}
